ui.views: move views, delete unused files
