+++ src/lando/ui/views/__init__.py
@@ -1,0 +1,3 @@
+from lando.ui.views import index, jobs, pull_requests, revisions
+
+__all__ = ["index", "jobs", "pull_requests", "revisions"]
