+++ src/lando/ui/views/base.py
@@ -1,0 +1,5 @@
+from django.views import View
+
+
+class LandoView(View):
+    pass
