+++ src/lando/ui/views/jobs.py
@@ -10,7 +10,7 @@ from lando.headless_api.models.automation_job import AutomationJob
 from lando.main.models import JobStatus, LandingJob, Worker, WorkerType
 from lando.main.models.jobs import BaseJob
 from lando.main.models.uplift import UpliftJob
-from lando.ui.views import LandoView
+from lando.ui.views.base import LandoView
 from lando.utils import treestatus
 
 logger = logging.getLogger(__name__)  # noqa: F821
